// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/catalog-product-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CatalogProductHttpResponse {
    alcoholPerBase?: number;
    barcode?: string | null;
    baseAmount?: number;
    baseUnit?: string;
    brand?: string | null;
    caloriesPerBase?: number;
    carbsPerBase?: number;
    category?: string | null;
    defaultPortionAmount?: number;
    description?: string | null;
    fatsPerBase?: number;
    fiberPerBase?: number;
    id?: string;
    imageUrl?: string | null;
    name?: string;
    productType?: string;
    proteinsPerBase?: number;
}