// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/catalog-recipe-export-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CatalogRecipeStepHttpResponse } from './catalog-recipe-step-http-response';

export interface CatalogRecipeExportHttpResponse {
    calculateNutritionAutomatically?: boolean;
    category?: string | null;
    cookTime?: number | null;
    description?: string | null;
    id?: string;
    imageUrl?: string | null;
    language?: string;
    languageConfirmed?: boolean;
    manualAlcohol?: number | null;
    manualCalories?: number | null;
    manualCarbs?: number | null;
    manualFats?: number | null;
    manualFiber?: number | null;
    manualProteins?: number | null;
    name?: string;
    prepTime?: number | null;
    servings?: number;
    steps?: Array<CatalogRecipeStepHttpResponse>;
}