// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/catalog-recipe-ingredient-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CatalogRecipeIngredientHttpResponse {
    amount?: number;
    amountText?: string | null;
    nestedRecipeId?: string | null;
    productId?: string | null;
    publicName?: string | null;
    publicUnit?: string | null;
    textName?: string | null;
}