// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/catalog-recipe-step-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CatalogRecipeIngredientHttpResponse } from './catalog-recipe-ingredient-http-response';

export interface CatalogRecipeStepHttpResponse {
    description?: string;
    imageAssetId?: string | null;
    imageAssetIds?: Array<string> | null;
    imageUrl?: string | null;
    ingredients?: Array<CatalogRecipeIngredientHttpResponse>;
    order?: number;
    title?: string | null;
}